CS50P final project/requirements.txt
python3
pytest
pygame-ce